Replace PlaceholdersDict with PlaceholderTranslations dataclass
This comments changes the Fetcher return from {key: List} (an assumes square) sequence-of-sequences + some associated metadata. This is the format natively returned by SQLAlchemy, among other things. It is always the format wanted by the default format applier. In the process, the Fetcher.fetch (and fetch_all) interface changes from required + optional placeholders to all-wanted-placeholders-in-desired-order + required

Multiple support methods for implementations are added to the Fetcher in the process.