Typing improvements
This adds a lot of "Higher-Kinded Type Vars" TODOs. Would like to define Translatable as something
like:

  Translatable = TypeVar("Translatable", Dict[NameType, List[IdType]])

The Translator.translate signature would then become:

  translate(t: Translatable[NameType, List[IdType]], ...) -> Translatable[NameType, List[Optional[str]]]

But that is not yet possible. I'd like the actual issue in the Python typing project, but I can't
do that without creating a link between them which I don't want.